+++ Functions/Get-LSS_CurrentConfigurationName.ps1
@@ -1,0 +1,13 @@
+Function Get-LSS_CurrentConfigurationName
+{
+  $RegBasePath = "HKLM:\SOFTWARE\PowerLSS"
+  $CurrentConfiguration = (Get-ItemProperty -Path $RegBasePath)."CurrentConfiguration"
+  if (!($CurrentConfiguration))
+  {
+    Write-Host "No current configuration has been set" -ForegroundColor Red
+  }
+  else
+  {
+    $CurrentConfiguration
+  }
+}