+++ Functions/Install-PowerLSS.ps1
@@ -6,17 +6,6 @@ Function Install-PowerLSS
 
   .DESCRIPTION
 	
-  .LINK
-      https://github.com/otabut/PowerLSS
-
-  .NOTES
-      Author: Olivier TABUT
-          0.3.0 release (28/01/2018)
-      ChangeLog: 
-          Initial version (14/01/2018)
-          0.2.0 release (21/01/2018)
-          0.3.0 release (28/01/2018)
-
   .PARAMETER Force
       Delete any previous scheduled task if exists before creating it
 
@@ -57,7 +46,10 @@ Function Install-PowerLSS
   Try
   {
     #Import PowerLSS helper module
-    Import-Module PowerLSS
+    If (!(Get-module PowerLSS))
+    {
+      Import-Module PowerLSS
+    }
 
     #Variables
     $Global:Log = @()
@@ -174,7 +166,7 @@ Function Install-PowerLSS
     if ($DoStart)
     {
       Write-LSS_Log -Step "Startup" -Status "Information" -Comment "Processing PowerLSS scheduled task startup"
-      $Command = Initialize-LSS_ScheduledTask
+      $Command = Start-LSS_ScheduledTask
       if ($Command.Result -eq 'Success')
       {
         Write-LSS_Log -Step "Startup" -Status "Information" -Comment "PowerLSS scheduled task started successfully"